// src/payload-types.ts
export interface User {
  id: number;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
// ...
export interface Experience {
  id: number;
  title: string;
  company: string;
  icon: string;
  job?: boolean | null;
  location?: string | null;
  href?: string | null;
  duration?: string | null;
  description?: string | null;
  tasks?:
    | {
        task: string;
        id?: string | null;
      }[]
    | null;
  badges?:
    | {
        badge: string;
        id?: string | null;
      }[]
    | null;
  createdBy?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
}